refactoring GoogleAPI class into old filename, updating transcription.py to match main HEAD
